+++ ml_project/tests/apis/predict_model_test.py
@@ -1,39 +1,10 @@
-import pandas as pd
-import pytest
-from sklearn.linear_model import LogisticRegression
-
 from clfit.apis import predict_model
 
 
-@pytest.fixture
-def features():
-    f = pd.DataFrame(
-        [
-            [0, 1, 0],
-            [1, 2, 1],
-        ]
-    )
-
-    return f
-
-
-@pytest.fixture
-def target():
-    return pd.Series([0, 1])
-
-
-@pytest.fixture
-def model(features, target):
-    model = LogisticRegression()
-    model.fit(features, target)
-
-    return model
-
-
-def test_train_model(model, features, target):
-    predicts = predict_model(model, features, return_proba=False)
+def test_train_model(trained_model, features, target):
+    predicts = predict_model(trained_model, features, return_proba=False)
 
     assert predicts.shape == target.shape
 
-    predicts = predict_model(model, features, return_proba=True)
+    predicts = predict_model(trained_model, features, return_proba=True)
     assert predicts.shape == target.shape
