+++ ml_project/tests/apis/train_model_test.py
@@ -1,33 +1,6 @@
-import pandas as pd
-import pytest
-from sklearn.linear_model import LogisticRegression
 from sklearn.utils.validation import check_is_fitted
 
 from clfit.apis import train_model
-
-
-@pytest.fixture
-def features():
-    f = pd.DataFrame(
-        [
-            [0, 1, 0],
-            [1, 2, 1],
-        ]
-    )
-
-    return f
-
-
-@pytest.fixture
-def target():
-    return pd.Series([0, 1])
-
-
-@pytest.fixture
-def model():
-    model = LogisticRegression()
-
-    return model
 
 
 def test_train_model(model, features, target):
